volatile only where a spin reads what the timer writes; the word helpers and the redundant casts are gone
**Three words are `volatile`, and nothing else is.** The game's units held
620 `volatile` tokens - on every `DG*` accessor, every struct and pointer
macro, every prototype, cast and local - and the qualifier buys one thing:
a loop that reads a word and does nothing else cannot have the read
hoisted. The game has three such loops, each on a word the timer thread
writes - the eight-tick spin on `DG44EE.frame_budget`,
`wait_and_latch_frame` on `DG5752.frame_flag`, `delay_five_ticks` on
`DG6430.ticks_left` - so those three fields carry it, with the reason
beside each, and the note above `DG8` says why the rest do not. The
timer thread is the port's own doing; where the two threads race on a
word (CLAUDE.md's concurrency entry) `volatile` never helped, and
everywhere it was asserting something untrue.

**What it was hiding.** With the reads visible to the optimiser,
`-Wmaybe-uninitialized` found the original's two uninitialised stack
reads. `open_sound_file`'s is settled: its three byte quads are the
`uint32_t found`, `uint32_t size` and `struct far_ptr cur` the code
computes with, `found` starts at the zero the search's own "nothing
matched" test is written for, and `size + 4` is one long where the
original carried by hand. `draw_part_selection`'s is not: the box is a
`struct extent16` (`width`, `height`) now, and at 0x16272 the original
halves the height before anything has set it - confirmed in the
disassembly - so the one warning left in the build is that line, and
the value it should read is still to be measured.

**`dg_rd16`/`dg_wr16`/`dg_rd32`/`dg_wr32` are retired.** All 228 sites
read `*(int16_t *)(p)` and write `*(int16_t *)(p) = v`: the host does
the unaligned load the original's `mov ax, [si+4]` did, and a
sanitizer's "misaligned" report on one describes the model.

**213 casts that said nothing are gone**: 79 to the expression's own
type, found by clang-tidy's readability-redundant-casting; 97 of the
converter's `(int16_t)(uint16_t)` and kin, whose inner half the outer
one repeats; 37 on literals that fit the variable they are assigned to.
A cast that truncates stays.

Built with the one warning above; the hybrid builds; `make test` passes
with all 34 solutions solving under simulation; the full verification -
the click-driven save against its reference, the briefing screen, all
34 solutions played for real - is recorded in the log beside this commit.

+++ reconstruct/src/sxovl_spkr.c
@@ -317,7 +317,7 @@ void sx_pitch_bend(uint16_t ax, uint16_t cx)
     uint8_t msb     = (uint8_t)cx;
     uint16_t value;
 
-    SXSPKR.word_033c = (int16_t)(uint16_t)(((uint16_t)msb << 7) | lsb);
+    SXSPKR.word_033c = (int16_t)(((uint16_t)msb << 7) | lsb);
 
     if (SXSPKR.byte_0348 != channel)
         return;
